recognition/requirements.txt
face_recognition==1.3.0
Pillow>=10.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
tqdm>=4.65.0
